requirements.txt
bencoder~=0.2.0
feedparser~=6.0.10
python-qbittorrent~=0.4.3
PyYAML~=6.0
requests~=2.28.1